fix broken tests around new metadata.py
